refactor: make interfaces folder
