+++ src/blar_graph/graph_construction/languages/javascript/javascript_parser.py
@@ -1,7 +1,9 @@
 import tree_sitter_languages
 
 from blar_graph.graph_construction.core.base_parser import BaseParser
-from blar_graph.graph_construction.utils.interfaces import GlobalGraphInfo
+from blar_graph.graph_construction.utils.interfaces.GlobalGraphInfo import (
+    GlobalGraphInfo,
+)
 
 
 class JavascriptParser(BaseParser):
