+++ src/blar_graph/graph_construction/languages/typescript/typescript_parser.py
@@ -5,7 +5,9 @@ from llama_index.packs.code_hierarchy.code_hierarchy import (
 )
 
 from blar_graph.graph_construction.core.base_parser import BaseParser
-from blar_graph.graph_construction.utils.interfaces import GlobalGraphInfo
+from blar_graph.graph_construction.utils.interfaces.GlobalGraphInfo import (
+    GlobalGraphInfo,
+)
 
 
 class TypescriptParser(BaseParser):
