+++ src/blar_graph/graph_construction/utils/interfaces/GlobalGraphInfo.py
